PolyUTS_source_code/requirements.txt
numpy
torch
torchvision
opencv-python
tqdm
pillow
